requirements.txt
flask
torch
torchvision
Pillow
numpy

requests
